# fix spc upload problems