--- a/bugs.md
+++ b/bugs.md
@@ -10,34 +10,22 @@ Generally, at least the following general issues seem at play:
 
 ## List
 
-- Actraiser 2
-  - Freezes after splash-screen (stuck in SPC upload)
 - Air Strike Patrol
   - Bugs due to not emulating mid-scanline effects
-- Battle Grand Prix
-  - Audio freezes after pressing start (SPC stuck in upload loop), game freezes at following SPC upload
 - Death Brade
   - Match ends instantly
 - Firepower 2000
   - Garbage line on title screen background
-- Hiouden - Mamono-tachi to no Chikai
-  - gets stuck on SPC upload on message after after Wolfteam-intro
 - Hook
   - Intros have incorrect colours and glitchy text
 - Mecarobot Golf
   - Some odd flashing of the ground in-game during movement
   - Sometimes freezes (?, hard to reproduce, happened once in lesson mode)
-- Illusion of Gaia
-  - Title screen is silent
-  - Freezes when pressing start (stuck in SPC upload)  
 - Jurassic Park
   - Sides of screen are not masked properly
 - Power Drive
   - Broken graphics on name select screen
   - Cannot control once in-game, only pause
-- Rendering Ranger R2
-  - Freezes on boot (stuck in SPC upload)
-- Soul Blazer
-  - Music sounds slightly off (missing drums on title intro)
 - Tales of Phantasia
-  - Sound effects and music sound off (wrong samples?)
+  - Voices sound fine until music starts playing, then everything gets garbled.
+    (probably unrelated to spc-upload -- investigating.)