… snes/spc.c