Hi @angelo-wf,
This pr fixes all of the spc transfer issues that I know of, the theory of why this is necessary is simple:
enix game jams data into the spc as fast as it possibly can, and running a single opcode at a time causes the following problem, example...  snes_catchupApu() needs to run 2 cycles, but the spc runs a whole opcode - of which would be 6 cycles long (0xfe).  The next spc-opcode is a mov from the port.
So, now the next write happens from cpu -> spc, and catching up apu will not run the spc at all since we are in debt by 4 cycles.  Now the spc "misses" the new data, and instead has the old data loaded from the port... uhoh :)

There is a new problem with "Tales of Phantasia".  The samples are no longer corrupted-sounding when they play individually, but when the game starts playing the music and the voice samples at the same time - everything gets garbled.  It sounds like there is a possible overrun in the sample buffer. (this is a guess).  I will investigate this soon,

best regards,
- dink