@dinkc64 2023-09-16
a little investigation into the sound issue of "Tales of Phantasia".  I noticed that the cycles per frame count go absolutely berzerk 78 frames into bootup for a few frames, then stays normal until the music+voice starts playing when you start a game.
from bootup:
apu / cpu cycles ran:  17088		357362 ... normal!
apu / cpu cycles ran:  17088		357368
apu / cpu cycles ran:  17088		357376
... cut ...
apu / cpu cycles ran:  17089		357378
apu / cpu cycles ran:  17086		357340
apu / cpu cycles ran:  17094		357366
apu / cpu cycles ran:  19836		414930  frame 78, starts to get abnormal
apu / cpu cycles ran:  25856		540760  .. even more abnormal
apu / cpu cycles ran:  23601		493572
apu / cpu cycles ran:  55338		1157266  "1157266" cycles per frame on the main cpu? yikes!
apu / cpu cycles ran:  23635		494340
apu / cpu cycles ran:  25889		541302
apu / cpu cycles ran:  13807		288880
apu / cpu cycles ran:  17087		357342 ... back to normal...
apu / cpu cycles ran:  17092		357362
apu / cpu cycles ran:  17086		357374


Here's a log from when the game starts:
apu / cpu cycles ran:  17090		357370   .. just the voice is playing here, no problem
apu / cpu cycles ran:  17086		357360
apu / cpu cycles ran:  17088		357366
apu / cpu cycles ran:  17089		357368
apu / cpu cycles ran:  17086		357362
apu / cpu cycles ran:  17089		357376
apu / cpu cycles ran:  17087		357358
apu / cpu cycles ran:  17088		357368
apu / cpu cycles ran:  17088		357370
apu / cpu cycles ran:  17088		357368
apu / cpu cycles ran:  17088		357358
apu / cpu cycles ran:  17254		360828  .. the game-screen appears, the music starts playing - it gets "warbley"/bad sounding.
apu / cpu cycles ran:  17085		357328
apu / cpu cycles ran:  16924		353934 .. notice the cycles oscillate between over-shooting and under-shooting?
apu / cpu cycles ran:  17253		360820
apu / cpu cycles ran:  16923		353916
apu / cpu cycles ran:  17254		360828
apu / cpu cycles ran:  16923		353926
apu / cpu cycles ran:  17149		358636
apu / cpu cycles ran:  17026		356078
apu / cpu cycles ran:  17255		360744
apu / cpu cycles ran:  16922		354006
apu / cpu cycles ran:  17551		367058
apu / cpu cycles ran:  16624		347656
apu / cpu cycles ran:  17555		367048
apu / cpu cycles ran:  16621		347696
apu / cpu cycles ran:  17552		367074
apu / cpu cycles ran:  16623		347640
apu / cpu cycles ran:  17558		367072
apu / cpu cycles ran:  16618		347658
apu / cpu cycles ran:  17554		367094
apu / cpu cycles ran:  16624		347642
apu / cpu cycles ran:  17549		367052
apu / cpu cycles ran:  16625		347676

I'm not too familiar with snes yet, so there is the possibility that this is normal(?)  What are your thoughts?

best regards,
- dink