@dinkc64 2023-09-17
to sum it up, the new sound-problem in "Tales of Phantasia" is caused by dma going past the end of frame.  I have some ideas on how to deal with this, will do some experimentation tonight!

best regards,
- dink