@dinkc64 2023-10-04
@angelo-wf, been doing tests over the past weeks and this PR is very solid.  The issue with "Tales of Phantasia" is due to the dma-frame-boundary issue. (I shouldn't have mentioned it in this PR, but I was unsure at the time...)

On to some really great news - my cx4 processor core is up and running, passes all of the tests, and runs Megaman X2/X3!  After a bit of clean-up I'll make a pr for it.

best regards,
- dink