@angelo-wf 2023-10-05
I want to look through this PR soon, I just haven't had much time for it lately. Nice that you got the CX4 working, looking forward to seeing how you implemented it.