@dinkc64 2023-10-11
Hi,
Just a little update on my cx4 core, so far things are looking good - I've spent the past week or so polishing it up, but there are a few caveats, here's a list...

simple/non-issue type things
~~
1: the core code (cx4.cpp) uses tab=4space, because my editor will not play nicely with tab=2spc, (it's something I'd been using since the mid 90's, go figure)
2: it's single context only
3: probably should playtest both games to the end :)

slightly difficult thing
~~
running the cx4 at its rated 20mhz, and the attract mode gameplay demo desychs (megaman gets killed, instead of megaman kills the boss).  though, running at half-speed (10mhz) causes everything to align perfectly.  I spent a solid week just trying to get to the bottom of this one, but, for now it eludes me.  I'll keep trying, though!

I would really like to do a playtest and try to get the timing 100% before making a PR, but if you would like to give it a try before then, I could e-mail over the files, or something like that :)

best regards,
- dink