@dinkc64 2023-10-25
Hi there, a little update on the CX4-core project:
...just fixing up some bugs found while play-testing, also took a short break to work on other things (hint, see new PR).  So far, things are looking good!

best regards,
- dink