@dinkc64 2023-11-04
Update:
While playtesting X2, found 3 gfx-related bugs:
1: A flicker on "You Got ...." scene (the screen after defeating boss), was caused by a slight timing inconsistency in my cx4 core
2: ? (I forgot the bug), due to unimpl. h/v latch in ppu, very easy to implement/fix.  snes.cpp 0x4201 & ppu.cpp 0x37 (ppu_read)
3: in Bubble Crab's stage, the fish ship's radar beam would flicker (on/off every other frame), finding the cause of this one was quite soul-crushing - but - it did allow me to get proper acquainted with the snes ppu!  Basically, the fix involves cancelling the halfColor shift if the clipmode == 1 or 2 logic hits (in ppu_handlePixel)

best regards,
- dink