@dinkc64 2023-11-08
A bit of awesome news:
One thing that kinda bothered me is the insane amount of cpu this project uses.  I spent perhaps a solid month attempting to optimize things to no avail, but...  finally, a couple of good ideas lead to super results today!
cut the cpu usage down from 10-11% of a core (13 being fully maxed, equal to 100%) to **7-9%!**
fastforward fps from 75-86 frames/per second to 92-104 frames/sec!! (ranges specified, as it fluxuates)
that's with the cx4 running (X2) :)

Hopefully I can get back to the playtesting so that I can finally wrap up and commit the CX4 project.

best regards,
- dink