@dinkc64 2023-11-14
stats with the latest on my fork - https://github.com/dinkc64/LakeSnes
game: Megaman X2 w/cx4 processor
fps values are with fast-forward (no frame limiting)
before optimizations: 76 - 94fps
after optimizations: 106 - 133fps
after optimizations & not drawing skipped frames**: 208 - 276fps
after optimizations + de-contexting ppu.c*:  118 - 153fps
after optimizations + de-contexting ppu.c & not drawing skipped frames**:  217 - 320fps ***success!***
** in ppu_runLine(), return after "if(!forcedBlank) ppu_evaluateSprites(line - 1);"
* de-contexting ppu.c involves moving the entire ppu structure into the ppu.c's global space as static variables.  

* and **, being experimental are not (yet) on git.

best regards,
- dink